server/src/requirements.txt
Django==1.9
djangorestframework==3.7.3
